% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog_tax}
\alias{eggnog_tax}
\title{Function for use with eggNOG emapper}
\usage{
eggnog_tax(df)
}
\arguments{
\item{df}{Dataframe downloaded from eggNOG emapper}
}
\description{
Extracts KO accession and gene name from emapper.annotations file. Appends phylogeny to dataframe
}
