% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGlevelsUniversal.R
\name{KOLevels}
\alias{KOLevels}
\title{Access KEGG Orthology}
\usage{
KOLevels(df, KOcolumnName, ECcolumnName = NULL)
}
\arguments{
\item{df}{Output csv from MPA}
}
\description{
Gets functional information for MPA output csv using the KEGG Orthology API.
}
