% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m5nr_join.R
\name{m5nr_join}
\alias{m5nr_join}
\title{Join function for MG-RAST}
\usage{
m5nr_join(dfKEGG, dfORG)
}
\arguments{
\item{dfKEGG}{Dataframe containing the Ontology-KO data}

\item{dfORG}{Dataframe containing the Organism data}
}
\description{
Joins both the "Organism" dataset and "Ontology-KO"
}
