% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m5nr_tax.R
\name{m5nr_tax}
\alias{m5nr_tax}
\title{Function to access Taxonomy}
\usage{
m5nr_tax(df)
}
\arguments{
\item{df}{Dataframe containing a column named "organism"}
}
\description{
Get phlyogeny from a given species through the UniProt API.
}
